+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [2.0.15]
 - Fix parent-query on sequence joins
+- Fix error handler on mutations when error happens on mutate call (before action is generated)
 
 ## [2.0.14]
 - Add `pc/mutation-data` helper